Fix URL parser, and handle no results scenario
The scraper is broken, and the URL format changed. So i fixed the
URL parser to the right ID position, and now specifically catch the
scenario where an empty DataFrame is returned.

I also removed funda_scraper as a requirement, and added it
as source code, and include it from there. Debatting whether to
include that source in the repo.

Currently only seems to work in python 3.10, no clue why